replaced CameraFollower with TransformFollower
